コンポーネントの分割する
